Remove default pw
If you setup a new host, you need to remember to create a dedicated
borgbackup_passphrase variable for this hosts. If you forget this,
this module will use yoursecret as password for the encryption key,
which is very unsecure.
People always do mistakes and we should prevent them to get into
situations like this.

Closes: #10